+++ tests/e2e/accessibility.spec.ts
@@ -1,5 +1,6 @@
 import { test, expect } from "@playwright/test";
-import { searchFor, runAxeAnalyze } from "./testUtils";
+import { mainPageSearch } from "./testUtils";
+import { runAxeAnalyze } from "./axeUtils";
 
 test("main page should have no accessibility violations (Axe)", async ({ page }, testInfo) => {
   await page.goto("/");
@@ -10,7 +11,7 @@ test("main page should have no accessibility violations (Axe)", async ({ page },
 
 test("search result with valid ID should have no accessibility violations (Axe)", async ({ page }, testInfo) => {
   // Use the helper function to search for a valid ID
-  await searchFor(page, "30056928150");
+  await mainPageSearch(page, "30056928150");
 
   // Verify that we're on the search results page
   await expect(page).toHaveURL("/melosys/sok");
@@ -25,7 +26,7 @@ test("search result with invalid ID should have no accessibility violations (Axe
   const invalidID = "INVALID123";
 
   // Use the helper function to search for the invalid ID
-  await searchFor(page, invalidID);
+  await mainPageSearch(page, invalidID);
 
   // Verify that we're on the search results page
   await expect(page).toHaveURL("/melosys/sok");
